Updated to Spring Boot 3.4.2
Updated to Spring Boot 3.4.2. Resolved all warning messages.

+++ src/main/java/com/bbs/entites/Message.java
@@ -1,16 +1,15 @@
 package com.bbs.entites;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import org.hibernate.annotations.Nationalized;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
@@ -20,7 +19,7 @@ public class Message {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private BigInteger id;
 	
 	@Column
 	private String title;
@@ -60,10 +59,10 @@ public class Message {
 		this.messageForum=messageForum;
 	}
 
-	public Integer getId() {
+	public BigInteger getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
 	public String getTitle() {
