+++ src/main/java/com/bbs/repos/MessageForumRepository.java
@@ -1,11 +1,13 @@
 package com.bbs.repos;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.bbs.entites.MessageForum;
 
 @Repository
-public interface MessageForumRepository extends JpaRepository<MessageForum,Integer> {
+public interface MessageForumRepository extends JpaRepository<MessageForum,BigInteger> {
 
 }
