+++ src/main/java/com/bbs/repos/MessageRepository.java
@@ -1,11 +1,13 @@
 package com.bbs.repos;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.bbs.entites.Message;
 
 @Repository
-public interface MessageRepository extends JpaRepository<Message, Integer> {
+public interface MessageRepository extends JpaRepository<Message, BigInteger> {
 
 }
