+++ src/main/java/com/bbs/services/MessageForumService.java
@@ -1,5 +1,6 @@
 package com.bbs.services;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,5 +10,5 @@ public interface MessageForumService {
 
 	public MessageForum save(MessageForum forum);
 	public List<MessageForum> findAll();
-	public Optional<MessageForum> findById(Integer id);
+	public Optional<MessageForum> findById(BigInteger id);
 }
