+++ src/main/java/com/bbs/services/MessageService.java
@@ -1,11 +1,12 @@
 package com.bbs.services;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import com.bbs.entites.Message;
 
 public interface MessageService {
 
-	public Optional<Message> findById(Integer id);
+	public Optional<Message> findById(BigInteger id);
 	public Message save(Message message);
 }
